requirements.txt
streamlit>=1.55.0
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pillow>=10.0.0
plotly>=5.0.0
